visualization/requirements.txt
kafka-python==2.0.2
numpy==1.24.3
pandas==2.0.3
dash==2.14.0
plotly==5.18.0